server/backend/requirements.txt
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
gpt4all==2.8.2
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
# Web- und Netzwerk-Bibliotheken
flask>=2.0.0
flask-cors>=3.0.0
websocket-client>=1.3.0
autobahn>=23.1.2
websockets>=10.4
twitchio>=2.6.0

# Umgebung und Konfiguration
python-dotenv>=0.21.0
typer>=0.7.0

# KI und Maschinenlernen
gpt4all>=1.0.0
transformers>=4.25.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
huggingface-hub>=0.17.0
openai>=1.5.0
google-generativeai>=0.3.0

# Systemüberwachung und -verwaltung
psutil>=5.9.0

# Utility-Bibliotheken
typing-extensions>=4.5.0
